fixed bug

+++ app/Http/Controllers/Api/AttractionsTakeTicketsTypesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Requests\Api\AttractionsTakeTicketsTypeRequest;
+use App\Models\Attraction;
 use App\Models\AttractionsTakeTicketsType;
 use App\Transformers\AttractionsTakeTicketsTypeTransformer;
 use Illuminate\Http\Request;
@@ -26,8 +27,6 @@ class AttractionsTakeTicketsTypesController extends Controller
 
     public function update(AttractionsTakeTicketsTypeRequest $request, AttractionsTakeTicketsType $attractionsTakeTicketsType)
     {
-        // todo...
-        // $this->authorize('update', $topic);
         $attractionsTakeTicketsType->update($request->all());
 
         return $this->response->item($attractionsTakeTicketsType, new AttractionsTakeTicketsTypeTransformer());
@@ -35,8 +34,10 @@ class AttractionsTakeTicketsTypesController extends Controller
 
     public function destroy(AttractionsTakeTicketsType $attractionsTakeTicketsType)
     {
-        // todo...
-        // $this->authorize('update', $topic);
+        $result = Attraction::where('take_tickets_type_id', $attractionsTakeTicketsType->id)->get();
+        if (!$result->isEmpty()) {
+            throw new \Dingo\Api\Exception\StoreResourceFailedException('请先删除该取票方式下边的景点, 再删除该分类');
+        }
 
         $attractionsTakeTicketsType->delete();
